[monaco] hook into keybinding resolution
to display proper keybinding in the quick command palette
